+++ src/compiler/prerender/test/prerender-index.spec.ts
@@ -1,6 +1,7 @@
 import * as d from '../../../declarations';
 import { doNotExpectFiles, expectFiles } from '../../../testing/utils';
-import { TestingCompiler, TestingConfig } from '../../../testing/index';
+import { TestingCompiler } from '../../../testing/testing-compiler';
+import { TestingConfig } from '../../../testing/testing-config';
 
 
 jest.setTimeout(10000);
@@ -10,6 +11,54 @@ describe('prerender index', () => {
   let c: TestingCompiler;
   let config: d.Config;
 
+
+  it('should pass properties down in prerendering', async () => {
+    config = new TestingConfig();
+    config.buildAppCore = true;
+    config.flags.prerender = true;
+
+    c = new TestingCompiler(config);
+    await c.fs.writeFile('/src/index.html', `
+      <script src="/build/app.js"></script>
+      <cmp-a></cmp-a>
+    `);
+    await c.fs.writeFile('/src/components/cmp-a/cmp-a.tsx', `
+      @Component({ tag: 'cmp-a' }) export class CmpA {
+
+        render() {
+          return <cmp-b someProp="property from parent" some-attr="attr from parent" somecrazy-ATTR="custom attr from parent" />;
+        }
+      }
+    `);
+    await c.fs.writeFile('/src/components/cmp-b/cmp-b.tsx', `
+      @Component({ tag: 'cmp-b' }) export class CmpB {
+        @Prop() someProp = 'unset';
+        @Prop() someAttr = 'unset';
+        @Prop({ attr: 'somecrazy-ATTR' }) someCustomAttr = 'unset';
+
+        render() {
+          return (
+            <div>
+              <p>{this.someProp}</p>
+              <p>{this.someAttr}</p>
+              <p>{this.someCustomAttr}</p>
+            </div>
+          );
+        }
+      }
+    `);
+    await c.fs.commit();
+
+    const r = await c.build();
+    expect(r.diagnostics).toEqual([]);
+
+    const index = await c.fs.readFile('/www/index.html');
+
+    expect(index).toContain('<p data-ssrc=\"1.0\"><!--s.1.0-->property from parent<!--/--> </p>');
+    expect(index).toContain('<p data-ssrc=\"1.1\"><!--s.1.0-->attr from parent<!--/--> </p>');
+    expect(index).toContain('<p data-ssrc=\"1.2\"><!--s.1.0-->custom attr from parent<!--/--> </p>');
+    expect(index).not.toContain('unset');
+  });
 
   it('should prerender w/ defaults', async () => {
     config = new TestingConfig();