+++ src/compiler/transpile/datacollection/test/prop-decorator.spec.ts
@@ -1,20 +1,16 @@
 import { gatherMetadata } from './test-utils';
 import { getPropDecoratorMeta } from '../prop-decorator';
 import { MEMBER_TYPE, PROP_TYPE } from '../../../../util/constants';
-import { mockConfig } from '../../../../testing/mocks';
 import * as path from 'path';
-import * as ts from 'typescript';
 
 
 describe('props decorator', () => {
-
-  const config = mockConfig();
 
   it('simple decorator', () => {
     let response;
     const sourceFilePath = path.resolve(__dirname, './fixtures/prop-example');
     const metadata = gatherMetadata(sourceFilePath, (checker, classNode, sourceFile, diagnostics) => {
-      response = getPropDecoratorMeta(config, checker, classNode, sourceFile, 'ClassName', diagnostics);
+      response = getPropDecoratorMeta(diagnostics, checker, classNode, sourceFile, 'ClassName');
     });
 
     expect(response).toEqual({
@@ -70,7 +66,56 @@ describe('props decorator', () => {
         propType: PROP_TYPE.Number,
         reflectToAttr: true
       },
+      width: {
+        attribName: 'width',
+        attribType: {
+          text: 'number',
+          typeReferences: {}
+        },
+        jsdoc: {
+          documentation: '',
+          name: 'width',
+          type: 'number',
+        },
+        memberType: 1,
+        propType: PROP_TYPE.Number,
+        reflectToAttr: false
+      },
     });
+  });
+
+  it('proper types', () => {
+    let response;
+    const sourceFilePath = path.resolve(__dirname, './fixtures/prop-types');
+    const metadata = gatherMetadata(sourceFilePath, (checker, classNode, sourceFile, diagnostics) => {
+      response = getPropDecoratorMeta(diagnostics, checker, classNode, sourceFile, 'ClassName');
+    });
+
+    // check strings
+    for (let i = 0; i < 11; i++) {
+      expect(response[`text${i}`].propType).toEqual(PROP_TYPE.String);
+    }
+
+    // number
+    for (let i = 0; i < 11; i++) {
+      expect(response[`nu${i}`].propType).toEqual(PROP_TYPE.Number);
+    }
+
+    // boolean
+    for (let i = 0; i < 4; i++) {
+      expect(response[`bool${i}`].propType).toEqual(PROP_TYPE.Boolean);
+    }
+
+    // TODO: revisit any vs unknown
+    // any
+    for (let i = 0; i < 5; i++) {
+      expect(response[`any${i}`].propType).toEqual(PROP_TYPE.Any);
+    }
+
+    // unknown
+    for (let i = 0; i < 2; i++) {
+      expect(response[`unknown${i}`].propType).toEqual(PROP_TYPE.Unknown);
+    }
   });
 
 });