+++ src/declarations/system.ts
@@ -8,11 +8,7 @@ export interface StencilSystem {
     typescriptVersion?: string;
     runtime?: string;
   };
-  createDom?(): {
-    parse(hydrateOptions: d.OutputTargetHydrate): Window;
-    serialize(): string;
-    destroy(): void;
-  };
+  createDom?(): CreateDom;
   createWatcher?(events: d.BuildEvents, paths: string, opts?: any): d.FsWatcher;
   generateContentHash?(content: string, length: number): string;
   fs?: d.FileSystem;
@@ -23,6 +19,7 @@ export interface StencilSystem {
   }): Promise<string[]>;
   isGlob?(str: string): boolean;
   loadConfigFile?(configPath: string): d.Config;
+  autoprefixCss?(input: string, opts?: any): Promise<string>;
   minifyCss?(input: string, opts?: any): {
     output: string;
     sourceMap?: any;
@@ -62,6 +59,13 @@ export interface StencilSystem {
     runInContext(code: string, contextifiedSandbox: any, options?: any): any;
   };
   workbox?: Workbox;
+}
+
+
+export interface CreateDom {
+  parse(hydrateOptions: d.OutputTargetHydrate): Window;
+  serialize(): string;
+  destroy(): void;
 }
 
 
@@ -122,6 +126,10 @@ export interface PackageJsonData {
   name?: string;
   version?: string;
   main?: string;
+  browser?: string;
+  module?: string;
+  'jsnext:main'?: string;
+  unpkg?: string;
   collection?: string;
   types?: string;
   files?: string[];