+++ src/server/platform-server.ts
@@ -2,7 +2,9 @@ import * as d from '../declarations';
 import { createDomApi } from '../renderer/dom-api';
 import { createQueueServer } from './queue-server';
 import { createRendererPatch } from '../renderer/vdom/patch';
-import { DEFAULT_STYLE_MODE, ENCAPSULATION, PROP_TYPE, RUNTIME_ERROR } from '../util/constants';
+import { DEFAULT_STYLE_MODE, ENCAPSULATION, RUNTIME_ERROR } from '../util/constants';
+import { enableEventListener } from '../core/listeners';
+import { fillCmpMetaFromConstructor } from './cmp-meta';
 import { getAppBuildDir } from '../compiler/app/app-file-naming';
 import { h } from '../renderer/vdom/h';
 import { initCoreComponentOnReady } from '../core/component-on-ready';
@@ -38,9 +40,8 @@ export function createPlatformServer(
 
   // initialize Core global object
   const Context: d.CoreContext = {};
-  Context.addListener = noop;
-  Context.enableListener = noop;
-  Context.emit = noop;
+  Context.enableListener = (instance, eventName, enabled, attachTo, passive) => enableEventListener(plt, instance, eventName, enabled, attachTo, passive);
+  Context.emit = (elm: Element, eventName: string, data: d.EventEmitterData) => domApi.$dispatchEvent(elm, Context.eventNameFn ? Context.eventNameFn(eventName) : eventName, data);
   Context.isClient = false;
   Context.isServer = true;
   Context.isPrerender = isPrerender;
@@ -77,14 +78,14 @@ export function createPlatformServer(
     attachStyles: noop,
     defineComponent,
     domApi,
-    emitEvent: noop,
+    emitEvent: Context.emit,
     getComponentMeta,
     getContextItem,
     isDefinedComponent,
     onError,
     nextId: () => config.namespace + (ids++),
     propConnect,
-    queue: createQueueServer(),
+    queue: (Context.queue = createQueueServer()),
     requestBundle: requestBundle,
     tmpDisconnected: false,
 
@@ -175,24 +176,7 @@ export function createPlatformServer(
           const componentConstructor = bundleExports[pascalCasedTagName];
 
           if (!cmpMeta.componentConstructor) {
-            // init component constructor
-            cmpMeta.componentConstructor = componentConstructor;
-
-            cmpMeta.membersMeta = {
-              'color': {}
-            };
-
-            if (cmpMeta.componentConstructor.properties) {
-              Object.keys(cmpMeta.componentConstructor.properties).forEach(memberName => {
-                const constructorProperty = cmpMeta.componentConstructor.properties[memberName];
-
-                if (constructorProperty.type) {
-                  cmpMeta.membersMeta[memberName] = {
-                    propType: PROP_TYPE.Any
-                  };
-                }
-              });
-            }
+            fillCmpMetaFromConstructor(componentConstructor, cmpMeta);
           }
 
           if (componentConstructor.style) {