+++ src/sys/node/node-sys.ts
@@ -54,8 +54,17 @@ export class NodeSystem implements d.StencilSystem {
     };
   }
 
+  private _existingDom: () => d.CreateDom;
+
   get createDom() {
+    if (this._existingDom) {
+      return this._existingDom;
+    }
     return createDom;
+  }
+
+  set createDom(val) {
+    this._existingDom = val;
   }
 
   createWatcher(events: d.BuildEvents, paths: string, opts: any) {
@@ -170,6 +179,33 @@ export class NodeSystem implements d.StencilSystem {
     }
 
     return config;
+  }
+
+  async autoprefixCss(input: string, opts: any): Promise<string> {
+    const modulePath = path.join(this.distDir, 'sys/node/auto-prefixer.js');
+    const module = require(modulePath);
+
+    const postcss = module.postcss;
+    const autoprefixer = module.autoprefixer;
+    if (typeof opts !== 'object') {
+      opts = {
+        browsers: [
+          'last 2 versions',
+          'iOS >= 8',
+          'Android >= 4.4',
+          'Explorer >= 11',
+          'ExplorerMobile >= 11'
+        ],
+        cascade: false,
+        remove: false
+      };
+    }
+    const prefixer = postcss([autoprefixer(opts)]);
+    const result = await prefixer.process(input, {
+      map: false,
+      from: undefined
+    });
+    return result.css;
   }
 
   minifyCss(input: string) {