+++ test/karma/stencil.config.js
@@ -1,0 +1,13 @@
+
+exports.config = {
+  outputTargets: [
+    {
+      type: 'www',
+      empty: false
+    }
+  ],
+  copy: [
+    { src: '**/*.html' }
+  ],
+  excludeSrc: []
+};